*Widget title saving in std::string. +Added Multiline mode in Label Widget. +Added background color in Label. +Small fixes in Label.cpp. *All size_t types seted to SHORT. +In Lable render function added special types filter. *CGUI_ERROR changed to 0xFFFF *Fix types bugs in TUI.cpp, Window.cpp. *Widget template constructor changed to (int,int,int,int,char const*). *Checked bug in Window::BitChecked(...) *Deleted all examples(((

+++ MicroTUI/TUI.cpp
@@ -22,7 +22,7 @@ namespace MicroTUI
 			if (!windowsList.empty())
 			{
 				//if (windowsList[windowsList.size()-1]->onFocus() == true)
-				for (int i = 0; i < windowsList.size(); i++)
+				for (int i = 0; i < (int)windowsList.size(); i++)
 						windowsList[i]->_Render_func(&__Screenbuffer);
 				//else
 				//{
